+++ tests/test_usage.py
@@ -13,6 +13,7 @@ def test_StatFin():
     db = statfin.StatFin()
     assert db.url == "https://statfin.stat.fi/PXWeb/api/v1/fi"
 
+
 def test_drilling():
     db = statfin.Vero()
     assert isinstance(db.Vero, statfin.PxWebAPI)
@@ -20,7 +21,8 @@ def test_drilling():
     assert isinstance(db.Vero.Henk.lopulliset, statfin.PxWebAPI)
     assert isinstance(db.Vero.Henk.lopulliset.tulot._101, statfin.Table)
     assert isinstance(db["Vero"]["Henk"].lopulliset["tulot"]._101, statfin.Table)
-    
+
+
 def test_variables():
     db = statfin.StatFin()
     tbl = db.StatFin.tyokay._115b
@@ -29,19 +31,18 @@ def test_variables():
     assert isinstance(tbl.Alue.SSS, statfin.Value)
     assert isinstance(tbl.Alue["SSS"], statfin.Value)
 
+
 def test_query():
     db = statfin.StatFin()
     tbl = db.StatFin.tyokay._115b
-    df = tbl.query(
-        {
-            "Alue": "SSS",  # Single value
-            "Pääasiallinen toiminta": "*",  # All values
-            "Sukupuoli": [1, 2],  # List of values
-            "Ikä": "18-64",  # Single value
-            "Vuosi": "2022",  # Single value
-            "Tiedot": "vaesto",  # Single value
-        }
-    )
+
+    q = tbl.query(Alue="SSS", Tiedot="vaesto")
+    q["Pääasiallinen toiminta"] = "*"  # All values
+    q.Vuosi = 2023  # Single value (will be cas to str)
+    q.Sukupuoli = [1, 2]  # Multiple values
+    q.Ikä = "18-64"  # Single value
+    df = q()
+
     assert isinstance(df, pd.DataFrame)
 
 
@@ -49,45 +50,25 @@ def test_cached_query():
     statfin.cache.clear()
     db = statfin.StatFin()
     tbl = db.StatFin.tyokay._115b
-    df = tbl.query(
-        {
-            "Alue": "SSS",
-            "Pääasiallinen toiminta": "*",
-            "Sukupuoli": [1, 2],
-            "Ikä": "18-64",
-            "Vuosi": "2022",
-            "Tiedot": "vaesto",
-        },
-        cache="test",
-    )
+
+    q = tbl.query(Alue="SSS", Tiedot="vaesto")
+    df = q("test")  # With cache id "test"
+
     assert isinstance(df, pd.DataFrame)
     assert os.path.isfile(".statfin_cache/test.df")
     assert os.path.isfile(".statfin_cache/test.meta")
 
-    df = tbl.query(
-        {
-            "Alue": "SSS",
-            "Pääasiallinen toiminta": "*",
-            "Sukupuoli": [1, 2],
-            "Ikä": "18-64",
-            "Vuosi": "2022",
-            "Tiedot": "vaesto",
-        },
-        cache="__test.cached.df",
-    )
+    df = q("test")
     assert isinstance(df, pd.DataFrame)
 
 
 def test_handles_comma_separator():
     db = statfin.StatFin()
     tbl = db.StatFin.ntp._11tj
-    df = tbl.query({
-        "Vuosineljännes": "*",
-        "Taloustoimi": "E2",
-        "Toimiala": "SSS",
-    })
-    assert(df.KAUSIT.notna().all())
-    assert(df.TASM.notna().all())
-    assert(df.TRENDI.notna().all())
-    assert(df.TYOP.notna().all())
 
+    df = tbl.query(Taloustoimi="E2", Toimiala="SSS")()
+
+    assert df.KAUSIT.notna().all()
+    assert df.TASM.notna().all()
+    assert df.TRENDI.notna().all()
+    assert df.TYOP.notna().all()